pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "automatedlbp-benchmarking"
version = "0.1.0"
description = "Automated LBP benchmarking pipeline"
requires-python = ">=3.9"
dependencies = [
  "numpy",
  "Pillow",
  "scikit-image",
]

[project.scripts]
lbp-benchmark = "automated_lbp_benchmarking.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]